add validate in servies

+++ services/attendance.services.ts
@@ -1,11 +1,25 @@
 import axios, { AxiosError } from "axios";
 import type { AttendanceCreateDto, AttendanceExportDto, AttendanceUpdateDto } from "~/interfaces/attendance.interface";
+import { checkToken } from "./auth"; 
 
 export const createAttendance = async (data: AttendanceCreateDto) => {
     try {
+        const authStore = useStore.authStore();
+
+        if (!await checkToken()) {
+            console.log('Unauthorize');
+
+            return null;
+        }
+
+        const apiUrl = useRuntimeConfig().public.apiUrl;
+
         const response = await axios({
             method: 'post',
-            url: "http://localhost/:8080/v1/attendance/create-attendance",
+            url: apiUrl + "attendance/create-attendance",
+            headers: {
+                'Authorization': 'Bearer ' + authStore.access_token
+            },
             data: {
                 attendance_date: data?.attendance_date,
                 student: data?.student,
@@ -18,15 +32,28 @@ export const createAttendance = async (data: AttendanceCreateDto) => {
         if (error instanceof AxiosError){
             return error.response?.data;
         }
-        return 'Something went wrong';
+        return null;
     }
 }
 
 export const updateAttendance = async (attendanceId: string, data: AttendanceUpdateDto) => {
     try {
+        const authStore = useStore.authStore();
+
+        if (!await checkToken()) {
+            console.log('Unauthorize');
+
+            return null;
+        }
+
+        const apiUrl = useRuntimeConfig().public.apiUrl;
+
         const response = await axios({
             method: 'put',
-            url: "http://localhost/:8080/v1/attendance/"+ attendanceId +"/update-attendance",
+            url: apiUrl + "attendance/"+ attendanceId +"/update-attendance",
+            headers: {
+                'Authorization': 'Bearer ' + authStore.access_token
+            },
             data: {
                 student: data.student
             }
@@ -37,15 +64,28 @@ export const updateAttendance = async (attendanceId: string, data: AttendanceUpd
         if (error instanceof AxiosError){
             return error.response?.data;
         }
-        return 'Something went wrong'
+        return null;
     }
 }
 
 export const deleteAttendance = async (attendanceId: string) => {
     try {
+        const authStore = useStore.authStore();
+
+        if (!await checkToken()) {
+            console.log('Unauthorize');
+
+            return null;
+        }
+
+        const apiUrl = useRuntimeConfig().public.apiUrl;
+
         const response = await axios({
             method: 'delete',
-            url: "http://localhost/:8080/v1/attendance/"+ attendanceId +"/delete-attendance"
+            url: apiUrl + "attendance/"+ attendanceId +"/delete-attendance",
+            headers: {
+                'Authorization': 'Bearer ' + authStore.access_token
+            }
         })
         return response.data;
 
@@ -53,15 +93,28 @@ export const deleteAttendance = async (attendanceId: string) => {
         if (error instanceof AxiosError){
             return error.response?.data;
         }
-        return 'Something went wrong';
+        return null;
     }
 }
 
 export const getAllAttendance = async (classroomId: string) => {
     try {
+        const authStore = useStore.authStore();
+
+        if (!await checkToken()) {
+            console.log('Unauthorize');
+
+            return null;
+        }
+
+        const apiUrl = useRuntimeConfig().public.apiUrl; 
+
         const response = await axios({
             method: 'get',
-            url: "http://localhost/:8080/v1/attendance/"+ classroomId +"/getAll-attendance"
+            url: apiUrl + "attendance/"+ classroomId +"/getAll-attendance",
+            headers: {
+                'Authorization': 'Bearer ' + authStore.access_token
+            }
         })
         return response.data;
         
@@ -69,15 +122,28 @@ export const getAllAttendance = async (classroomId: string) => {
         if (error instanceof AxiosError){
             return error.response?.data;
         }
-        return 'Something went wrong';
+        return null;
     }
 }
 
 export const getAttendanceById = async (attendanceId: string) => {
     try {
+        const authStore = useStore.authStore();
+
+        if (!await checkToken()) {
+            console.log('Unauthorize');
+
+            return null;
+        }
+
+        const apiUrl = useRuntimeConfig().public.apiUrl; 
+
         const response = await axios({
             method: 'get',
-            url: "http://localhost/:8080/v1/attendance/"+ attendanceId +"/getById-attendance"
+            url: apiUrl + "attendance/"+ attendanceId +"/getById-attendance",
+            headers: {
+                'Authorization': 'Bearer ' + authStore.access_token
+            }
         })
         return response.data;
 
@@ -85,15 +151,28 @@ export const getAttendanceById = async (attendanceId: string) => {
         if (error instanceof AxiosError){
             return error.response?.data;
         }
-        return 'Something went wrong';
+        return null;
     }
 }
 
 export const exportAttendance = async (data: AttendanceExportDto) => {
     try {
+        const authStore = useStore.authStore();
+
+        if (!await checkToken()) {
+            console.log('Unauthorize');
+
+            return null;
+        }
+
+        const apiUrl = useRuntimeConfig().public.apiUrl; 
+         
         const response = await axios({
             method: 'get',
-            url: "http://localhost:8080/v1/attendance/export-attendance-excel",
+            url: apiUrl + "attendance/export-attendance-excel",
+            headers: {
+                'Authorization': 'Bearer ' + authStore.access_token
+            },
             data: {
                 classroom_id: data.classroom_id
             }
@@ -105,7 +184,7 @@ export const exportAttendance = async (data: AttendanceExportDto) => {
             return error.response?.data;
         
         }
-        return 'Something went wrong';
+        return null;
     }
 }
